requirements.txt
gitpython
requests
pyyaml
pyinstaller